New DL run
New DL run with deep and wide and conv model. For the first 30 epochs, use cosine annealing for the lr, then decrease the lr by 1/1.1 each epoch. The run lasted 56 epochs and took 7612 seconds. The best validation loss was 0.42197 and the corrsponding auc was 0.88563. This was worse than the run with the deep and wide model, but only marginally.