New DL run
Save info, model and prediction of new DL run as ..._best. In the run the ReLU activations were replaced with GELU activations. The run started with a lr of 1e-3 and used Cosine Annealing to reduce the lr to 1e-6 by epoch 30. After epoch 30, the lr was reduced by a factor of 1/1.1 each epoch. The run lasted a total of 56epochs and lasted 6787seconds. The best validation loss was 0.41826 at epoch 30 and the corresponding auc was 0.88781.