New DL run
New DL run. In the run the last activation was changed from Tanh to GELU. The run started with a lr of 1e-3 and used Cosine Annealing to reduce the lr to 1e-6 by epoch 30. After epoch 30, the lr was reduced by a factor of 1/1.1 each epoch. The run lasted a total of 56 epochs and lasted 6787seconds. The best validation loss was 0.41849 at epoch 51 and the corresponding auc was 0.88769.