New DL run
Save info, model and prediction of new DL run as ..._best. In the run the GELU activations were used, but the last activation was Tanh. The run started with a lr of 1e-3 and used Cosine Annealing to reduce the lr to 1e-8 by epoch 50. The run lasted a total of 53 epochs and lasted 7472 seconds. The best validation loss was 0.41312 at epoch 48 and the corresponding auc was 0.89105.