New DL run
Save info, model and prediction of new DL run as ..._best. GELU activations were used and 2 more hidden layers were added to the deep model. The run started with a lr of 1e-3 and used Cosine Annealing to reduce the lr to 1e-8 by epoch 50. The run lasted a total of 51 epochs. The best validation loss was 0.41098 at epoch 51 and the corresponding auc was 0.89222. The loss was still decreasing so training for more epochs would almost certainly improve the model.