New HIGGS DL run
Save info, model and prediction of new DL run as ..._best. GELU activations were used except for the last layer which used Tanh. The run started with a lr of 1e-3 and used Cosine Annealing to reduce the lr to 1e-10 by epoch 100. The run lasted a total of 121 epochs and lasted 21658 seconds. The best validation loss was 0.40821 at epoch 97 and the corresponding auc was 0.89391.