New SUSY DL run
save prediction and infos for a SUSY DL run as ..._best. The run used 7 GELU layers followed by a Tanh layer. The lr started at 1e-3 and cosine annealing was used to get the lr down to 1e-7 by epoch 50. After epoch 50 the lr stayed constant. The best validation loss was 0.42085 in epoch 72, the corresponding auc was 0.87987. The run laster for 82 epochs and took 6661 seconds.