New HIGGS DL run (training info not saved, this needs to be repeated)
New HIGGS DL run.  Only the predictions are saved, the training infos were lost. This run needs to be repeated to also save the training infos.

For the training set, the final loss was 0.39354 and the auc score 0.90154. For the validation set the final loss was 0.409414 and the auc score was 0.89333.

The deep model used 2**11 units and 8 dense blocks. The first 7 dense blocks used GELU activation, the 8th dense block used a tanh activation. A dropout rate of 0.2 and weight decay of 0.05 were used.

The learning rate started at 1e-3 and was reduced to 1e-5 via cosine annealing by epoch 100, after which it did not change.

The run lasted for 115 epochs and took 19600 seconds. The best epoch was epoch 106.