New HIGGS DL run
New HIGGS DL run.

The results for the test set were: 0.39350 loss, 0.90156 auc, for the validation set: 0.409397 loss, 0.89334 auc

the model used 8 dense blocks, the first 7 with GELU activation, the last with Tanh. 2**11 units were used and a dropout rate of 0.2 and 0.05 weight decay.

The lr started at 1e-3 and decreased to 1e-5 by epoch 100 with cosine annealing, after which the lr stayed constant.

The run lasted 115 epochs, the best epoch was epoch 105. The run took 21297 seconds.